decimal to integer

+++ db/migrate/20150116154638_add_total_cost_to_participants.rb
@@ -1,5 +1,5 @@
 class AddTotalCostToParticipants < ActiveRecord::Migration
   def change
-    add_column :participants, :total_cost, :decimal
+    add_column :participants, :total_cost, :integer
   end
 end
